partial skip_defaults implementation
Running into trouble getting skip_defaults passed correctly. I’m not understanding the full flow of the logic from the route decorators to the response serialization.

This is a Starlette class which has include_in_schema but does not have skip_defaults - how do we handle this? Do we need to?

    # From routing.py - line 331
    elif isinstance(route, routing.Route):
        self.add_route(
            prefix + route.path,
            route.endpoint,
            methods=route.methods,
            include_in_schema=route.include_in_schema,
            name=route.name,
        )

Not completely sure I'm handling passing skip_defaults through get_app correctly.

Haven't done any websocket work - not sure if websocket route needs to be handled for this (probably so)