Only use default parameter values for path parameters (#450)
This change allows endpoint functions to be defined with parameters
with defaults, but for those defaults to be ignored if the function
parameter subsequently appears as a path parameter. The default
value is only operative if the parameter is a query parameter.

This change includes a new set of tests for this functionality.