Fix #259: generate OpenAPI docs for responses with no content.
This is a backwards-compatible change, as far as I know, for responses that explicitly return content.  There are two basic changes:

* FastAPI now knows which response codes do not return bodies, and suppresses the `content` element in their API docs.
* If the response_class for a route has no media type, FastAPI assumes that it will not return content.  (If this route specifies additional responses that do return content, those responses are assumed to return `application/json` content and use an appropriate response class.)

This change introduces two new test files which check the new logic and keep the code coverage at 100%.

This change also updates the tutorial documentation to reflect the new functionality.