WIP: Populate summary and description of an endpoint from its docstring
This commit changes how the OpenAPI summary and description for an
endpoint are populated from the docstring of the function implementing
the endpoint.

Old behavior: Only the description is populated from the docstring

New behavior: Both the summary and the description are populated from
the docstring unless explicitly provided as decorator arguments

The new behavior is more aligned with PEP 257 specifying that a
docstring consists of summary line and an extended description, neatly
mapping to the corresponding fields in OpenAPI specification.