Cache cloning of pydantic types globally for 2+x init speed-up
If `response_model=...` is specified in a route handler, FastAPI
creates a copy of the type to avoid accidental leakage of
secrets. This is done in a pure way, but is not cached. This means
that that if there's models that occur several times (for instance,
multiple routes that return the same model), identical work is done
and re-done. This change adds a cache for the conversion: when
converting a type that's already been seen, just use the cache.

This dramatically improves the start-up time of the server, making for
a nicer developer experience, and reducing cold-start times if run in
FaaS/serverless platforms like AWS Lambda.

For instance, we have an app with 40 routes, defined in the usual way
with an `app = FastAPI(...)` in app.py and various routers across
other files.

I measured the impact on this change in two ways.

- Just the time to initialise all the routes etc (command below):
- The time to reach "Application startup complete" when actually
  running (`uvicorn.run(..., port=8000, reload=True, debug=True,
  workers=1)`)

| measurement    | 0.70.0 (s) | this patch (s) |
|----------------|-----------:|---------------:|
| initialisation |       2.60 |           1.26 |
| full startup   |       5.94 |           2.15 |

Initialisation measurement command:

    python -c 'import time; start = time.perf_counter(); import ....app; print(time.perf_counter() - start)'