Update benchmarks.md
@DongwonGim    
I have accepted your feedback. Thankyou 🚀

@tiangolo
I hope you merge this document.
There are many documents still open in #2017. 
I'm curious as to why it can't be merged. 
Even if it's not perfect, if you merge it, it will be easier for people to find parts to fix on the FastAPI webpage, so please merge it.