Docs: fix incorrect example for dicts in 3.6+
As part of the new python 3.9+ examples there's been a small error in the example for dicts. Instead of adding a new example, the old code was replaced and linked to be used for both 3.6+ and 3.9+.

This merely restores the old example for 3.6 and moves the new example to a new file with the "_py39" suffix (so it's in line with the other similar examples) and fixes the link in the markdown.