Add echo to live command to inform that the server is running.
Echo in serve command don't work for me, if run live command.
I decided to add the output of the message, because after running
the command, nothing is displayed in the console,
and this confused me for the first time.
Therefore, IMHO, it will be better to deduce some information message
that the server was started.