deleting a route from APIRouter
There are some scenarios which someone may need to delete a route from APIRouter. For example you may use a thirdparty library which provides a router for you. But you donot want to use all of the provided endpoints. In this case it is useful to remove some of endpoints