Sort methods to stabalize operation id generation
The methods attribute accessed by the `generate_unique_id` function is modeled as a python set, which is not ordered and between runs might change. This results in the same fastapi generating different openapi specs if started multiple times.

This also relates to the bigger issue https://github.com/tiangolo/fastapi/issues/4740 but is not a solution for that issue.