feat: automatically add localised language name to language selector
The order of the content of the  "Languages" and "alternate" sections in the mkdocs.yaml files have changed, which impacts the order the languages are shown on the UI in the language selector. Previously English was always shown first and the other languages were alphabetised by language code, now English is no longer a special case and appears in the middle of the list.

The mapping of language code to local language name was sourced from https://quickref.me/iso-639-1. Where were discrepancies in capitalisation and/or in language name the existing names were reused.