Introduce a gate/check GHA job
This adds a GHA job that reliably determines if all the required
dependencies have succeeded or not.

It also allows to reduce the list of required branch protection CI
statuses to just one — `check`. This reduces the maintenance burden
by a lot and have been battle-tested across a small bunch of projects
in its action form and in-house implementations of other people.

It is now in use in aiohttp (and other aio-libs projects), CherryPy,
attrs, coveragepy, conda, pydantic, some of the Ansible repositories,
open edX, pip-tools, spaceship-prompt, all of the jaraco's projects
(like `setuptools`, `importlib_metadata`), some PyCQA, PyCA, PyPA and
pytest projects, a few AWS Labs projects.

The story behind this is explained in more detail at
https://github.com/marketplace/actions/alls-green#why.