Add security utility to make and check nonces.
HTTP digest access authentication requires the use of a nonce value to
prevent replay attacks. This commit introduces a simple stateless nonce
for that purpose, following the format recommended by RFC7616 (section
3.3) with a ':' delimited string of timestamp in miliseconds, resource,
and server secret data. Note, the nonce utility hash-and-truncates the
secret and request URL path to limit the size of the nonce to 48 chars.
Also note, the request URL path is used to describe the resource as
opposed to the recommended ETag for compatibility reasons.