Update HTTPDigest to be consistent with RFC 7616
This is an initial implementation of HTTP Digest Access Authentication
consistent to [RFC 7616](https://datatracker.ietf.org/doc/html/rfc7616).
As the RFC only provides guidance, and does not enforce all elements of
the standard, this commit only implements the crucial aspects to ensure
the challenge-response negotiation is at least usable. Notably,
non-quoted usernames, and proxy-authentication is not yet supported by
this change. Furthermore, this commit uses a stateless-Nonce generation
as to avoid overcomplicating the design, future changes may wish to
address this to allow servers to validate the nonce is indeed one-time
usage.