Cherry-picked master branch updates
I do it this way, so that I can keep reordering commits (doing it via the GitHub "sync fork" button breaks this feature)

Note that, because I do it like this, the GitHub GUI will assume that this branch is out of sync with the master branch. But that's not the case, this branch is in sync.