Remove `./` from start of internal links
Includes only the links where i just removed the `./` from the start and did nothing else.

I checked this and the links in the next commits with a script, see `linkchecker.py` in the `scripts/` commit in my workbranch (#10854). That script checks if local links (including images and python code files) are valid and well-formed. It also suggests hash parts of links that can possibly be translated.