Normalize info boxes
These boxes are often misunderstood in translations. I also did translate the keywords in the beginning.

The syntax is "!!! " followed by <keyword>, optionally followed by " " <translation>

<keyword> is required and is one of "info|note|tip|check|warning|danger". It defines the coloring of the info box. When it was missing, I added it, checking the related English .md.

<translation> may come after <keyword> and is inside of "". If the translation was available, I added it.

A global search under docs/ in VS Code with the regex "!!! (?!tip|warning|info|danger|note|check)" shows info boxes which do not comply to this scheme.

I left the invalid boxes under docs/fa/unchanged, because this is a right to left language and I am unsure how to write it there.