fix: allow_inf_nan option for Param subclasses
There was a typo in the __init__ method of the Param class which
misnamed the allow_inf_nan option in the kwargs passed to the
underlying Pydantic FieldInfo initializer.  This resulted in the
validator failing to enforce the constraint, and hence inf and nan
values being passed to the route handler where they should not have.