📝 Fix async test example not to trigger DeprecationWarning
The code sample in `docs_src/async_tests/test_main.py` triggers the
following deprecation warning with recent versions of `httpx`:

    DeprecationWarning: The 'app' shortcut is now deprecated.
    Use the explicit style 'transport=ASGITransport(app=...)' instead.

We fix it as suggested by the warning.

Note that due to https://github.com/encode/httpx/issues/3111 the new way
of creating `AsyncClient` passes type checks only when using httpx
version 0.27.2 or newer. This is not a problem for FastAPI itself though
because it does not type check code samples.