Update first-steps.md
I changed this

*path operation decorator* --> *path operation* decorator

and

**"path operation decorator"** --> **"path operation" decorator"**

because the "path operation" part is the specification of the current decorator, and not the full name of this feature. I didn't want new readers to call it "path operation decorator" but rather "decorator".

I don't know what you think about this small change. You'll let me know.