Update #13929
There are a few more changes here than necessary, because I created these translations from scratch without an existing translation and without wordlists in the German LLM, to learn which words the AI actually needs to have defined (less than expected). I then was too lazy to remove the not super necessary changes, :-D. But The translations are good, I manually checked them.

You may apply this commit on #13929. Regenerating it is not necessary.