Add space between ellipsis and words
Also added rule to prompt.

I thought it is okay, meanwhile, in German when there is no space inbetween, (which is why i did not fix these changes before) but this is actually not the case. In English both spellings are okay.

Has one change in the English document. If we leave those backticks, the LLM should by definition not translate it, but we want this here.