Improve instructions for existing translations
I noticed that it forgot to remove a removed sentence and to add an added part of a sentence, when I reordered/updated a few sentences in the test file. It was stubborn even when retranslating. After these reformulations and the the added sentence at the end it saw these changes and properly synced them.

This is also cleanup for the abbr issue, not directly related.